# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MaskChain API
  version: 1.0.0
  description: API Gateway с Content Shield (AI DLP) и управлением профилями политик безопасности
  license:
    name: Proprietary
    identifier: Proprietary
servers:
  - url: /api/v1
    description: API v1
security: []
paths:
  /chat/completions:
    post:
      operationId: createChatCompletion
      summary: Прокси к LLM провайдеру (chat completion)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                model:
                  type: string
                  example: gpt-4
                stream:
                  type: boolean
                  default: false
                messages:
                  type: array
      responses:
        '200':
          description: Ответ от провайдера (прозрачный proxy)
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
...
  /shield/mask:
    post:
      operationId: maskText
      summary: Маскировать PII в тексте
      parameters:
        - name: mask_id
          in: query
          schema:
            type: string
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
          description: Текст с замаскированными данными
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /shield/unmask:
    post:
      operationId: unmaskText
      summary: Восстановить замаскированный текст
      parameters:
        - name: mask_ids
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  schemas:
    ErrorInfo:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
    ApiResponse:
      type: object
      properties:
        data:
          oneOf:
            - type: object
            - type: array
            - type: 'null'
        error:
          oneOf:
            - $ref: '#/components/schemas/ErrorInfo'
            - type: 'null'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
  responses:
    SuccessResponse:
      description: Успешный ответ в едином envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
...
    ErrorResponse:
      description: Ошибка в едином envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'